# Conflicts:
#	Next2/src/Next2/App.xaml.cs
#	Next2/src/Next2/Models/API/DTO/CustomerModelDTO.cs
#	Next2/src/Next2/Models/CustomerBindableModel.cs
#	Next2/src/Next2/Models/FullOrderBindableModel.cs
#	Next2/src/Next2/Models/GiftCardModel.cs
#	Next2/src/Next2/Models/PaidOrderBindableModel.cs
#	Next2/src/Next2/Services/Mock/CustomersMock.cs
#	Next2/src/Next2/Services/Mock/MockService.cs
#	Next2/src/Next2/ViewModels/CustomersViewModel.cs
#	Next2/src/Next2/ViewModels/Dialogs/AddGiftCardDialogViewModel.cs
#	Next2/src/Next2/ViewModels/Mobile/InputGiftCardPageViewModel.cs
#	Next2/src/Next2/ViewModels/PaymentCompleteViewModel.cs

+++ Next2/src/Next2/Models/API/DTO/GiftCardModelDTO.cs
@@ -11,7 +11,7 @@ namespace Next2.Models.API.DTO
 
         public DateTime Expire { get; set; }
 
-        public double TotalBalance { get; set; }
+        public decimal TotalBalance { get; set; }
 
         public Guid? CustomerId { get; set; }
     }
