+++ Next2/src/Next2/Models/OrderBindableModel.cs
@@ -22,8 +22,8 @@ namespace Next2.Models
 
         public string OrderNumberText { get; set; } = string.Empty;
 
-        public float Total { get; set; }
+        public decimal Total { get; set; }
 
-        public float PriceTax { get; set; }
+        public decimal PriceTax { get; set; }
     }
 }
